ML-42 add meta models

+++ synthesized/metadata_new/value/continuous.py
@@ -2,21 +2,21 @@ from typing import Optional
 
 import numpy as np
 
-from .base import Domain, Scale, Ring
+from ..base import Domain, Scale, Ring
 
 
-class Integer(Scale[np.int64]):
+class Integer(Scale[int]):
     class_name: str = 'Integer'
     dtype: str = 'int64'
 
     def __init__(
             self, name: str, domain: Optional[Domain[np.int64]] = None, nan_freq: Optional[float] = None,
-            min: Optional[np.int64] = None, max: Optional[np.int64] = None
+            min: Optional[int] = None, max: Optional[int] = None
     ):
         super().__init__(name=name, domain=domain, nan_freq=nan_freq, min=min, max=max)
 
 
-class Float(Ring[np.float64]):
+class Float(Ring[float]):
     class_name: str = 'Float'
     dtype: str = 'float64'
